[Fix] Support multiple ports with LACP fallback
Fixes #7555 and #8866.

While the LACP fallback HLD correctly describes the behavior for LAGs with
multiple ports, the actual implementation only worked for single-port LAGs.

When multiple ports are present, the fallback mode would select all ports,
therefore making it useless for the purpose described in the HLD (PXE-booting).

Therefore, this commit adds a patch to implement the desired behavior:
If all non-disabled ports are defaulted, *one* port goes into fallback mode and
gets selected for the LAG. As soon as any port receives a LACPDU, the fallback
port goes back to defaulted.
The fallback port is selected by higher port priority or, if both have equal
priority (the default), by lower port number (as described in the HLD).

Signed-off-by: Lukas Stockner <lstockner@genesiscloud.com>