Use debian's mirrors without chaining mirrors
`apt-get update` will return a non-zero exit code with at least one
invalid mirror, even if there are valid mirrors specified. Therefore
given a non-zero exit breaks a command sequence, it is not possible
to specify redundant mirrors with the expectation they'll failover.

This is relevant because `apt-get` is chained in sequence together
in the templated `Dockerfile`'s present for creating the Debian-based
build environments.